+++ RELEASE_NOTES.md
@@ -22,6 +22,7 @@
 - Added logs in src/wq/ql/notify.py (#656)
 - Added logs in src/wq/core/task.py (#662)
 - Replaced entity with getter (#652)
+- Moved env vars to cfgreader.py (#678)
 
 ### Bug fixes